JERSEY-2126: Actually disable chunked encoding when requested
Even though in ApacheHttpClient4Handler.getHttpEntity(ClientRequest) a
BufferedHttpEntity is created if the property PROPERTY_CHUNKED_ENCODING_SIZE
is set to null, the BufferedHttpEntity will never buffer the contents
because it uses EntityUtils.toByteArray(HttpEntity) internally which
requests the data from HttpEntity.getContent(). Which returns null.

So, don't return null.